requirements.txt
flask
google-generativeai
langchain
langchain_community
faiss-cpu
langchain_google_genai
